feat(api-client): updates collection page layout
