+++ packages/api-client/src/views/Collection/CollectionNavigation.vue
@@ -6,6 +6,8 @@ import { RouterLink, useRouter, type RouteLocationNamedRaw } from 'vue-router'
 import { PathId } from '@/routes'
 import { useActiveEntities } from '@/store'
 
+import CollectionInfoForm from './CollectionInfoForm.vue'
+
 const { currentRoute } = useRouter()
 const { activeCollection } = useActiveEntities()
 
@@ -93,27 +95,27 @@ const routes = computed<CollectionSidebarEntry[]>(() => [
 ])
 </script>
 <template>
-  <!-- <div
-    class="border-b min-h-11 -mb-0.25 flex items-center px-2.5 text-sm font-medium sticky top-0 bg-b-1 xl:rounded-none">
-    <span class="text-c-2">
-      {{ activeCollection?.info?.title || 'Untitled Collection' }}
-    </span>
-  </div> -->
-  <div
-    class="bg-b-1 sticky top-0 flex min-h-11 items-center gap-5 border-b px-3 text-sm font-medium">
-    <RouterLink
-      v-for="({ to, displayName }, i) in routes"
-      :key="i"
-      class="flex h-full cursor-pointer items-center whitespace-nowrap border-b border-solid border-transparent text-center text-sm font-medium no-underline has-[:focus-visible]:outline"
-      :class="
-        typeof to.name === 'string' &&
-        typeof currentRoute.name === 'string' &&
-        currentRoute.name?.startsWith(to.name)
-          ? 'text-c-1 border-white'
-          : 'text-c-2 hover:text-c-1'
-      "
-      :to="to">
-      {{ displayName }}
-    </RouterLink>
+  <div class="mx-auto w-full md:max-w-[50dvw]">
+    <CollectionInfoForm />
+    <div
+      class="gap-1/2 bg-b-1 sticky top-0 flex h-fit items-center text-sm font-medium">
+      <RouterLink
+        v-for="({ to, displayName }, i) in routes"
+        :key="i"
+        class="-ml-0.5 flex h-11 cursor-pointer items-center whitespace-nowrap px-2 text-center text-sm font-medium no-underline -outline-offset-1 has-[:focus-visible]:outline"
+        :to="to">
+        <span
+          class="flex-center h-full w-full border-b border-transparent"
+          :class="
+            typeof to.name === 'string' &&
+            typeof currentRoute.name === 'string' &&
+            currentRoute.name?.startsWith(to.name)
+              ? 'text-c-1 border-white'
+              : 'text-c-2 hover:text-c-1'
+          ">
+          {{ displayName }}
+        </span>
+      </RouterLink>
+    </div>
   </div>
 </template>
