+++ packages/api-client/src/views/Collection/CollectionSettings.vue
@@ -1,7 +1,4 @@
 <script setup lang="ts">
-import DeleteSidebarListElement from '@/components/Sidebar/Actions/DeleteSidebarListElement.vue'
-import { PathId } from '@/router'
-import { useActiveEntities, useWorkspace } from '@/store'
 import {
   ScalarButton,
   ScalarIcon,
@@ -10,6 +7,10 @@ import {
   useModal,
 } from '@scalar/components'
 import { useRouter } from 'vue-router'
+
+import DeleteSidebarListElement from '@/components/Sidebar/Actions/DeleteSidebarListElement.vue'
+import { PathId } from '@/router'
+import { useActiveEntities, useWorkspace } from '@/store'
 
 const { activeCollection, activeWorkspace, activeWorkspaceCollections } =
   useActiveEntities()
@@ -58,16 +59,16 @@ function handleDeleteCollection() {
 
 <template>
   <div
-    class="flex flex-col gap-12 h-full mx-auto py-8 w-full md:max-h-[82dvh] md:max-w-[50dvw]">
+    class="mx-auto flex h-full w-full flex-col gap-12 px-1.5 py-8 md:max-h-[82dvh] md:max-w-[50dvw]">
     <div class="flex flex-col gap-4">
       Features
       <!-- Watch Mode -->
-      <div class="bg-b-2 border rounded-lg text-sm">
+      <div class="bg-b-2 rounded-lg border text-sm">
         <div
-          class="bg-b-1 border flex gap-4 items-center justify-between -m-1/2 p-3 rounded-t-lg">
+          class="bg-b-1 -m-1/2 flex items-center justify-between gap-4 rounded-t-lg border p-3">
           <div>
             <h3>Watch Mode</h3>
-            <p class="mt-1 text-c-2">
+            <p class="text-c-2 mt-1">
               When enabled, the OpenAPI document will be polled for changes. The
               collection will be updated automatically.
             </p>
@@ -79,23 +80,23 @@ function handleDeleteCollection() {
             @update:modelValue="handleToggleWatchMode" />
         </div>
         <div
-          class="flex items-center overflow-x-auto py-1.5 whitespace-nowrap text-c-1">
+          class="text-c-1 flex items-center overflow-x-auto whitespace-nowrap py-1.5">
           <div class="flex items-center">
             <template v-if="activeCollection?.documentUrl">
               <span class="bg-b-2 sticky left-0 pl-3 pr-2">Source</span>
               <a
-                class="group text-c-2 no-underline rounded hover:underline pr-3"
+                class="text-c-2 group rounded pr-3 no-underline hover:underline"
                 :href="activeCollection.documentUrl"
                 target="_blank">
                 {{ activeCollection.documentUrl }}
                 <ScalarIcon
-                  class="hidden group-hover:inline ml-1 w-2.5"
+                  class="ml-1 hidden w-2.5 group-hover:inline"
                   icon="ExternalLink" />
               </a>
             </template>
             <template v-else>
               <ScalarIcon
-                class="ml-3 mr-2 w-4 text-c-2"
+                class="text-c-2 ml-3 mr-2 w-4"
                 icon="NotAllowed"
                 size="sm" />
               <span class="text-c-2 pr-3">
@@ -112,16 +113,16 @@ function handleDeleteCollection() {
     <div class="flex flex-col gap-4">
       <div>Danger Zone</div>
       <div
-        class="border flex items-center justify-between p-3 rounded-lg text-sm">
+        class="flex items-center justify-between rounded-lg border p-3 text-sm">
         <div>
           <h3>Delete Collection</h3>
-          <p class="mt-1 text-c-2">
+          <p class="text-c-2 mt-1">
             Be careful, my friend. Once deleted, there is no way to recover the
             collection.
           </p>
         </div>
         <ScalarButton
-          class="gap-1.5 font-medium px-2.5 h-8 shadow-none focus:outline-none custom-scroll whitespace-nowrap"
+          class="custom-scroll h-8 gap-1.5 whitespace-nowrap px-2.5 font-medium shadow-none focus:outline-none"
           :variant="'danger'"
           @click="deleteModal.show()">
           Delete Collection
